+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fix typo #2
 - custom `gitonic` path -> file `~/.gitonic/thonnycontrib.cfg` 
 - changed to subprocess implementation
+  - platform windows specfic popen and shlex handling
 -